requirements.txt
pandas==1.5.3
plotly==5.13.0
python-binance==1.0.16
tabulate==0.9.0